kkjj-13 <feature>(<Air>) Air 등록, 조회 / TODO : 수정/삭제, 필터/정렬

+++ src/main/java/renewal/awesome_travel_backoffice/air/dto/response/SeatClassResponseDto.java
@@ -9,8 +9,8 @@ import renewal.awesome_travel_backoffice.air.utiles.SeatClassType;
 public class SeatClassResponseDto {
     private Long id;
     private SeatClassType classType;
-    private Integer price;
-    private Integer maxSeats;
-    private Integer availableSeats;
+    private Long price;
+    private Long maxSeats;
+    private Long availableSeats;
 }
